clean_data remove, Forwardfill => Backfill good
In the original preprocessing() function, stocks are effectively delisted if they have a single nan value because dropna(axis = 1). Later on we forward-fill then back-fill which does nothing because there are no nan values, since we delisted the stocks that have them. nans are generally scarce across time, so just applying the original fills should prove to be sufficient, instead of removing stocks because they have a single nan over a 20-year period.

Alternatively one could set dropna(axis = 0), but then we remove time-steps which means we have jumps in time which breaks the stochastic control assumption of evenly space time-steps.